# Add fast grouping method for CategoricalArray keys