--- a/test/dataframerow.jl
+++ b/test/dataframerow.jl
@@ -45,7 +45,7 @@ module TestDataFrameRow
     @test hash(DataFrameRow(df, 2)) != hash(DataFrameRow(df, 6))
 
     # check that hashrows() function generates the same hashes as DataFrameRow
-    df_rowhashes, _ = DataFrames.hashrows(df, false)
+    df_rowhashes, _ = DataFrames.hashrows(tuple(DataFrames.columns(df)...), false)
     @test df_rowhashes == [hash(dr) for dr in eachrow(df)]
 
     # test incompatible frames