--- a/test/grouping.jl
+++ b/test/grouping.jl
@@ -222,9 +222,11 @@ module TestGrouping
                        Key2 = CategoricalArray(["A", "B", "A", "B"]),
                        Value = 1:4)
         gd = groupby(df, :Key1)
+        @test length(gd) == 2
         @test gd[1] == DataFrame(Key1=["A", "A"], Key2=["A", "B"], Value=1:2)
         @test gd[2] == DataFrame(Key1=["B", "B"], Key2=["A", "B"], Value=3:4)
         gd = groupby(df, [:Key1, :Key2])
+        @test length(gd) == 4
         @test gd[1] == DataFrame(Key1="A", Key2="A", Value=1)
         @test gd[2] == DataFrame(Key1="A", Key2="B", Value=2)
         @test gd[3] == DataFrame(Key1="B", Key2="A", Value=3)
@@ -233,13 +235,24 @@ module TestGrouping
         levels!(df[:Key1], ["Z", "B", "A"])
         levels!(df[:Key2], ["Z", "B", "A"])
         gd = groupby(df, :Key1)
-        @test gd[1] == DataFrame(Key1=["A", "A"], Key2=["A", "B"], Value=1:2)
-        @test gd[2] == DataFrame(Key1=["B", "B"], Key2=["A", "B"], Value=3:4)
+        @test length(gd) == 2
+        @test gd[1] == DataFrame(Key1=["B", "B"], Key2=["A", "B"], Value=3:4)
+        @test gd[2] == DataFrame(Key1=["A", "A"], Key2=["A", "B"], Value=1:2)
         gd = groupby(df, [:Key1, :Key2])
+        @test length(gd) == 4
         @test gd[1] == DataFrame(Key1="A", Key2="A", Value=1)
         @test gd[2] == DataFrame(Key1="A", Key2="B", Value=2)
         @test gd[3] == DataFrame(Key1="B", Key2="A", Value=3)
         @test gd[4] == DataFrame(Key1="B", Key2="B", Value=4)
+        # Make first level unused too
+        df.Key1[1:2] .= "B"
+        gd = groupby(df, :Key1)
+        @test length(gd) == 1
+        @test gd[1] == DataFrame(Key1="B", Key2=["A", "B", "A", "B"], Value=1:4)
+        gd = groupby(df, [:Key1, :Key2])
+        @test length(gd) == 2
+        @test gd[1] == DataFrame(Key1="B", Key2="A", Value=[1, 3])
+        @test gd[2] == DataFrame(Key1="B", Key2="B", Value=[2, 4])
 
         # Check that CategoricalArray column is preserved when returning a value...
         res = combine(d -> DataFrame(x=d[1, :Key2]), groupby(df, :Key1))
@@ -337,19 +350,34 @@ module TestGrouping
         @test res.x isa Vector{Int}
     end
 
-    @testset "grouping with missings" begin
-        global df = DataFrame(Key1 = ["A", missing, "B", "B", "A"],
-                              Key2 = CategoricalArray(["B", "A", "A", missing, "A"]),
-                              Value = 1:5)
+    @testset "grouping with missings" for df in
+        (DataFrame(Key1 = ["A", missing, "B", "B", "A"],
+                   Key2 = ["B", "A", "A", missing, "A"],
+                   Value = 1:5),
+         DataFrame(Key1 = categorical(["A", missing, "B", "B", "A"]),
+                  Key2 = ["B", "A", "A", missing, "A"],
+                  Value = 1:5),
+         DataFrame(Key1 = ["A", missing, "B", "B", "A"],
+                  Key2 = categorical(["B", "A", "A", missing, "A"]),
+                  Value = 1:5),
+         DataFrame(Key1 = categorical(["A", missing, "B", "B", "A"]),
+                  Key2 = categorical(["B", "A", "A", missing, "A"]),
+                  Value = 1:5))
 
         @testset "sort=false, skipmissing=false" begin
-            global gd = groupby(df, :Key1)
+            gd = groupby(df, :Key1)
             @test length(gd) == 3
-            @test gd[1] == DataFrame(Key1=["A", "A"], Key2=["B", "A"], Value=[1, 5])
-            @test gd[2] ≅ DataFrame(Key1=missing, Key2="A", Value=2)
-            @test gd[3] ≅ DataFrame(Key1=["B", "B"], Key2=["A", missing], Value=3:4)
+            if df.Key1 isa CategoricalVector # Order of missing changes
+                @test gd[1] == DataFrame(Key1=["A", "A"], Key2=["B", "A"], Value=[1, 5])
+                @test gd[2] ≅ DataFrame(Key1=["B", "B"], Key2=["A", missing], Value=3:4)
+                @test gd[3] ≅ DataFrame(Key1=missing, Key2="A", Value=2)
+            else
+                @test gd[1] == DataFrame(Key1=["A", "A"], Key2=["B", "A"], Value=[1, 5])
+                @test gd[2] ≅ DataFrame(Key1=missing, Key2="A", Value=2)
+                @test gd[3] ≅ DataFrame(Key1=["B", "B"], Key2=["A", missing], Value=3:4)
+            end
 
-            global gd = groupby(df, [:Key1, :Key2])
+            gd = groupby(df, [:Key1, :Key2])
             @test length(gd) == 5
             @test gd[1] == DataFrame(Key1="A", Key2="B", Value=1)
             @test gd[2] ≅ DataFrame(Key1=missing, Key2="A", Value=2)
@@ -359,12 +387,12 @@ module TestGrouping
         end
 
         @testset "sort=false, skipmissing=true" begin
-            global gd = groupby(df, :Key1, skipmissing=true)
+            gd = groupby(df, :Key1, skipmissing=true)
             @test length(gd) == 2
             @test gd[1] == DataFrame(Key1=["A", "A"], Key2=["B", "A"], Value=[1, 5])
             @test gd[2] ≅ DataFrame(Key1=["B", "B"], Key2=["A", missing], Value=3:4)
 
-            global gd = groupby(df, [:Key1, :Key2], skipmissing=true)
+            gd = groupby(df, [:Key1, :Key2], skipmissing=true)
             @test length(gd) == 3
             @test gd[1] == DataFrame(Key1="A", Key2="B", Value=1)
             @test gd[2] == DataFrame(Key1="B", Key2="A", Value=3)
@@ -372,13 +400,13 @@ module TestGrouping
         end
 
         @testset "sort=true, skipmissing=false" begin
-            global gd = groupby(df, :Key1, sort=true)
+            gd = groupby(df, :Key1, sort=true)
             @test length(gd) == 3
             @test gd[1] == DataFrame(Key1=["A", "A"], Key2=["B", "A"], Value=[1, 5])
             @test gd[2] ≅ DataFrame(Key1=["B", "B"], Key2=["A", missing], Value=3:4)
             @test gd[3] ≅ DataFrame(Key1=missing, Key2="A", Value=2)
 
-            global gd = groupby(df, [:Key1, :Key2], sort=true)
+            gd = groupby(df, [:Key1, :Key2], sort=true)
             @test length(gd) == 5
             @test gd[1] ≅ DataFrame(Key1="A", Key2="A", Value=5)
             @test gd[2] == DataFrame(Key1="A", Key2="B", Value=1)
@@ -387,13 +415,13 @@ module TestGrouping
             @test gd[5] ≅ DataFrame(Key1=missing, Key2="A", Value=2)
         end
 
-        @testset "sort=false, skipmissing=true" begin
-            global gd = groupby(df, :Key1, sort=true, skipmissing=true)
+        @testset "sort=true, skipmissing=true" begin
+            gd = groupby(df, :Key1, sort=true, skipmissing=true)
             @test length(gd) == 2
             @test gd[1] == DataFrame(Key1=["A", "A"], Key2=["B", "A"], Value=[1, 5])
             @test gd[2] ≅ DataFrame(Key1=["B", "B"], Key2=["A", missing], Value=3:4)
 
-            global gd = groupby(df, [:Key1, :Key2], sort=true, skipmissing=true)
+            gd = groupby(df, [:Key1, :Key2], sort=true, skipmissing=true)
             @test length(gd) == 3
             @test gd[1] == DataFrame(Key1="A", Key2="A", Value=5)
             @test gd[2] == DataFrame(Key1="A", Key2="B", Value=1)
@@ -402,7 +430,7 @@ module TestGrouping
     end
 
     @testset "iteration protocol" begin
-        global gd = groupby(DataFrame(A = [:A, :A, :B, :B], B = 1:4), :A)
+        gd = groupby(DataFrame(A = [:A, :A, :B, :B], B = 1:4), :A)
         for v in gd
             @test size(v) == (2,2)
         end