When grouping on a single `CategoricalArray` column, we don't need a `Dict` to get group codes.
This approach cannot be used when joining, as the `Dict` is needed.
`group_rows` now handles the sorting, which allows skipping this step if the grouping is guaranteed to be sorted (as is the case for `CategoricalArray`).
This mechanism can be extended in the future to handle multiple `CategoricalArray` keys, or integer keys (using radix sort).

This makes grouping about 4 times faster on simple benchmarks, even with the optimized code from https://github.com/JuliaData/DataFrames.jl/pull/1565 (which BTW is still used when grouping on several `CategoricalArray` keys). Of course grouping only represents a small part of the time taken by a split-apply-combine operation, so typically the win isn't that big.
```julia
using DataFrames, BenchmarkTools
df = DataFrame(rand(500_000, 10));
df.key = categorical(string.(rand(1:100, 500_000)));

# Master
julia> @btime groupby(df, :key);
  8.258 ms (51 allocations: 15.45 MiB)

# This PR
julia> @btime groupby(df, :key);
  2.284 ms (64 allocations: 7.64 MiB)

df = DataFrame(rand(500_000, 10));
df.key = categorical(string.(rand(1:10_000, 500_000)));

# Master
julia> @btime groupby(df, :key);
  27.451 ms (54 allocations: 15.68 MiB)

# This PR
julia> @btime groupby(df, :key);
  7.595 ms (69 allocations: 7.86 MiB)
```