@bkamins 2018-11-16
Thank you. Looks good. One thing that might be added is to internally document the meaning of return value of `row_group_slots` somewhere for future developers (in a comment probably is enough).