@bkamins 2018-11-16
Thank you. I think we are good to merge.

Do you have any benchmark how we stand now in the benchmark against other frameworks?