@nalimilan 2018-11-16
> Do you have any benchmark how we stand now in the benchmark against other frameworks?

Not yet. I'd like to implement the type-stable `combine` here or in DataFramesMeta first.