@nalimilan 2018-11-16
Though it shouldn't make a big difference for https://github.com/h2oai/db-benchmark/issues/30 when there are enough rows per group. So we could ask them to update the benchmark and run it again. But `SubDataFrame` deprecations will probably slow down the simplest syntax, so better drop them first.