@bkamins 2018-11-16
They rerun them automatically on the latest release of benchmarked packages and they are updated regularly AFAIK, so we simply have to move forward with the package I guess 😄.